puzzles/iot-twitter/requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.8.24
chardet==3.0.4
colorama==0.4.0
future==0.16.0
idna==2.7
more-itertools==4.3.0
numpy==1.15.2
oauthlib==2.1.0
pandas==0.23.4
pbr==4.3.0
pluggy==0.7.1
py==1.6.0
PySocks==1.6.8
pytest==3.8.2
python-dateutil==2.7.3
python-twitter==3.4.2
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.23
